move mongoose schemas to separate files
